Refactor Fader to be an Observer of Subtractor
This is now working for the master gain, but there are a lot of TODOs

- Lint and tests are all failing
- Use native javascript `get` and `set` syntax for gain, will need to
refactor all other params to follow this
- Observer & Observable are exports of Observe.js. Subtractor inherits
from Observable, but Fader cannot inherit from Observer because it
already is a subclass of HTMLElement. For now I want to leave those
classes there as a reference of the pattern sued
