+++ src/components/Fader.js
@@ -2,17 +2,19 @@ import { pointToPercent } from '../utils/maths'
 import styles from '../sass/fader.scss';
 
 class Fader extends HTMLElement {
-  static get observedAttributes() {
-    return ['id', 'name', 'min', 'max', 'value',]
-  }
-
   connectedCallback() {
+    // if observable and bind attributes are preset, register this element as an observer
+    this.observable = eval(this.getAttribute('observe'))
+    this.bind = this.getAttribute('bind')
+    if (this.observable && this.bind) {
+      this.observable.registerObserver(this)
+    }
+    
     this.id = this.getAttribute('id')
     this.name = this.getAttribute('name')
     this.min = this.getAttribute('min')
     this.max = this.getAttribute('max')
     this.value = this.getAttribute('value')
-    this.oninput = new Function('value', this.getAttribute('oninput'))
     this.template = `
       <label for="fader__input" class="fader" id="fader">
         <input class="fader__input" id="fader__input" type="range" 
@@ -73,26 +75,26 @@ class Fader extends HTMLElement {
 
     this.faderInput.addEventListener('input', (e) => {
       const inputValue = parseInt(e.target.value)
-      const inputMax = parseInt(e.target.max)
-      const inputMin = parseInt(e.target.min) 
-      
-      const percent = pointToPercent(inputMin, inputMax, inputValue)
-      const top = this.maxTop * (1 - percent)
-      
-      this.faderKnob.style.top = `${parseInt(top)}px`
-      this.faderValue.innerText = inputValue
-
-      // call the oninput function passed in as an HTML attribute
-      this.oninput(parseInt(inputValue))
+      this.setTop(inputValue)
+      this.faderValue.innerText = parseInt(inputValue)
+      this.observable[this.bind] = parseInt(inputValue)
     })
+  }
 
-    this.faderInput.addEventListener('change', (e) => {
-      const inputValue = parseInt(e.target.value)
-      this.faderValue.innerText = inputValue
-    })
+  notify(observable) {
+    this.faderInput.value = this.observable[this.bind]
+    this.faderValue.innerText = parseInt(this.observable[this.bind])
+    this.setTop(this.observable[this.bind])
+  }
 
-    // trigger input event so `top` style is set in DOM
-    this.faderInput.dispatchEvent(new Event('input'))
+  setTop(inputValue) {
+    const inputMax = parseInt(this.faderInput.max)
+    const inputMin = parseInt(this.faderInput.min) 
+    
+    const percent = pointToPercent(inputMin, inputMax, inputValue)
+    const top = this.maxTop * (1 - percent)
+    
+    this.faderKnob.style.top = `${parseInt(top)}px`
   }
 
   mousemove (_this, x, y, oldValue, e) {
