+++ src/presets/Init.json
@@ -1,34 +1,34 @@
 {
- "name": "",
- "author": "",
- "description": "",
- "settings": {
-  "master": {
-   "gain": 0.5
-  },
-  "super": {
-   "polyphony": 1,
-   "detune": 0
-  },
-  "osc1": {
-   "enabled": 1,
-   "waveform": 3,
-   "octave": 0,
-   "semi": 0,
-   "cent": 0
-  },
-  "osc2": {
-   "enabled": 0,
-   "waveform": 3,
-   "octave": 0,
-   "semi": 0,
-   "cent": 0
-  },
-  "filter1": {
-   "type": "lowpass",
-   "frequency": 22050,
-   "q": 0.10000000149011612,
-   "gain": 0
+  "name": "Init",
+  "author": "",
+  "description": "",
+  "settings": {
+    "master": {
+      "gain": 50
+    },
+    "super": {
+      "polyphony": 1,
+      "detune": 0
+    },
+    "osc1": {
+      "enabled": 1,
+      "waveform": 1,
+      "octave": 0,
+      "semi": 0,
+      "cent": 0
+    },
+    "osc2": {
+      "enabled": 1,
+      "waveform": 1,
+      "octave": 0,
+      "semi": 0,
+      "cent": 0
+    },
+    "filter1": {
+      "type": "lowpass",
+      "frequency": 22050,
+      "q": 0.10000000149011612,
+      "gain": 0
+    }
   }
- }
 }
